splitting up io.py (#51)
* copy header part

* moving set_tags

* moving stuff

* moving more

* moving moving

* done moving

* cleaning up
