+++ src/riogrande/io.py
@@ -43,20 +43,20 @@ from ._helper import (
     view_to_window,
     get_scale_factor,
 )
-# is_needed
-# needs_work (module should be moved, needs docs)
-# is_tested (at least partially)
-# usedin_both
 
+
+# TODO: Adapt this to rioG (or similar)
 # this is our namespace for tags
 NS = 'LANDIV'
+
 
 # TODO: General Idea - maybe we can merge some of these into io_.py class structure - so we avoid having both.
 #  --> yet it is nice to have the function by themselves as well without direct need of class structures
 
-
 def set_tags(src, bidx:int|None=None, ns:str=NS, **tags):
-    # TODO: is_needed - needs_work - is_tested - usedin_both
+    # is_needed
+    # needs_work (should be made internal?)
+    # is_tested
     """Update tags for a dataset or a single band of a dataset.
 
     Since metadata in a tif file is stored as a string the value of each tag is
@@ -102,10 +102,6 @@ def set_tags(src, bidx:int|None=None, ns:str=NS, **tags):
       The value provided is converted to a string with `helper.serialize`
       before the tag is written to the file.
     """
-    # is_needed
-    # needs_work (should be made internal?)
-    # is_tested
-    # usedin_both (is internal to IO)
 
     if bidx is None:
         bidx = 0
@@ -114,34 +110,34 @@ def set_tags(src, bidx:int|None=None, ns:str=NS, **tags):
     src.update_tags(ns=ns, bidx=bidx, **serialized_tags)
 
 def get_tags(src, bidx:int|None=None, ns:str=NS):
-    # TODO: is_needed - needs_work - is_tested - usedin_both
-    """Get all the tags and deserialize the values
+        # is_needed
+        # needs_work (should be internal)
+        # is_tested
+        """Get all the tags and deserialize the values
 
-    Parameters
-    ----------
-    src:
-      `tif` file openend with `rasterio.open`
-    bidx:
-      Index of the band to get tags from (starting from 1 as is the convention
-      in rasterio). If set to `None` then the tags for the entire dataset are
-      returned.
-    ns:
-      The namespace to get the tags from.
-      
-      ..Note::
-        It is dicouraged to change this value from the default as all tagging
-        related methods of this package use the same default namespace.
-    """
-    # is_needed
-    # needs_work (should be internal)
-    # is_tested
-    # usedin_both (is internal to IO)
-    if bidx is None:
-        bidx = 0  # get the tags for the files metadata
-    return deserialize(src.tags(bidx=bidx, ns=ns))
+        Parameters
+        ----------
+        src:
+          `tif` file openend with `rasterio.open`
+        bidx:
+          Index of the band to get tags from (starting from 1 as is the convention
+          in rasterio). If set to `None` then the tags for the entire dataset are
+          returned.
+        ns:
+          The namespace to get the tags from.
+          
+          ..Note::
+            It is dicouraged to change this value from the default as all tagging
+            related methods of this package use the same default namespace.
+        """
+        if bidx is None:
+            bidx = 0  # get the tags for the files metadata
+        return deserialize(src.tags(bidx=bidx, ns=ns))
 
 def find_bidxs(src, ns:str=NS, **tags):
-    # TODO: is_needed - needs_work - is_tested - usedin_both
+    # is_needed
+    # neews_work (should be internal)
+    # not_tested
     """Find all bands in src for which all tags match
 
     Parameters
@@ -158,10 +154,6 @@ def find_bidxs(src, ns:str=NS, **tags):
       Arbitrary number of keyword arguments that will be compared to the tags
       of the bands in the dataset.
     """
-    # is_needed
-    # neews_work (should be internal)
-    # not_tested
-    # usedin_both (internal to IO)
 
     _tags = sanitize(tags) 
     matching_bidxs = []
@@ -172,7 +164,9 @@ def find_bidxs(src, ns:str=NS, **tags):
     return matching_bidxs
 
 def get_bidx(src, ns:str=NS, **tags)->None|int:
-    # TODO: is_needed - needs_work - is_tested - usedin_both
+    # is_needed
+    # needs_work
+    # not_tested (used in tests)
     """Get the index of the band with matching tags
 
     ..Note::
@@ -237,10 +231,6 @@ def get_bidx(src, ns:str=NS, **tags)->None|int:
       Arbitrary number of keyword arguments that will be compared to the tags
       of the bands in the dataset.
     """
-    # is_needed
-    # no_work
-    # not_tested (used in tests)
-    # usedin_both
 
     if 'indexes' in tags or not tags:
         bidx = tags.get('indexes', 1)  # return 1 if nothing is provided
@@ -261,7 +251,9 @@ def get_bidx(src, ns:str=NS, **tags)->None|int:
     return bidx
 
 def get_bands(source:str, ns:str=NS, **tags)->list[tuple[str,int]]:
-    # TODO: is_needed - needs_work - is_tested - usedin_both
+    # is_needed
+    # needs_work
+    # not_tested (used in tests)
     """Find all bands that match specific tags
 
     This method check the metadata (including those of bands)
@@ -293,10 +285,6 @@ def get_bands(source:str, ns:str=NS, **tags)->list[tuple[str,int]]:
       Arbitrary number of keyword arguments that will be compared to the tags
       of each tif file
     """
-    # is_needed
-    # no_work
-    # not_tested (used in tests)
-    # usedin_both
 
     _tags = sanitize(tags)
     _sources = glob.glob(source)
@@ -313,7 +301,9 @@ def get_bands(source:str, ns:str=NS, **tags)->list[tuple[str,int]]:
 
 
 def load_map(source:str, **tags)->dict:
-    # TODO: is_needed (but only in tests)
+    # is_needed (this is only used in tests)
+    # needs_work (replace usage with `load_block` and get rid of it)
+    # not_tested (used in tests)
     """Load a specific band from a tif file
 
     See `load_block` for details
@@ -324,18 +314,15 @@ def load_map(source:str, **tags)->dict:
        Returns the callback of
        `load_block(source=source, view=None, scaling_params=None, **tags)`
     """
-    # is_needed (this is only used in tests)
-    # needs_work (replace usage with `load_block` and get rid of it)
-    # not_tested (used in tests)
-    # usedin_both
     return load_block(source=source, view=None, scaling_params=None, **tags)
-
 
 def load_block(source:str,
                view:None|tuple[int,int,int,int]=None,
                scaling_params:dict|None=None,
                **tags)->dict:
-    # TODO: is_needed - needs_work - not_tested - usedin_both
+    # is_needed
+    # needs_work
+    # is_tested
     """Get a block from a specific band of a *.tif file along with the transform
 
     You can select what band(s) to load by passing keyword arguments as tags
@@ -373,10 +360,6 @@ def load_block(source:str,
        orig_meta: The meta information of the original .tif file
        orig_profile: The profile information of the original .tif file
     """
-    # is_needed
-    # no_work
-    # is_tested
-    # usedin_both
     window=view_to_window(view)
     with rasterio.open(source) as img:
         # TODO: rasterio Window allows using slices. In doing so we could
@@ -421,13 +404,14 @@ def load_block(source:str,
             'orig_profile': img.profile.copy()
         }
 
-
 def write_band(src:DatasetWriter,
                data:NDArray,
                bidx:int=1,
                window:Window|None=None,
                **tags):
-    # TODO: is_needed (only 1 use) - needs_work - is_tested - usedin_both
+    # is_needed
+    # needs_work
+    # not_tested
     """Write data to a specific band of a tif file and set the tags
 
 
@@ -444,19 +428,16 @@ def write_band(src:DatasetWriter,
       The value provided is converted to a string with `helper.serialize`
       before the tag is written to the file.
     """
-    # is_needed
-    # no_work
-    # not_tested
-    # usedin_processing
     src.write(data, indexes=bidx, window=window)
     set_tags(src, bidx=bidx, **tags)
-
 
 def update_band(src:DatasetWriter,
                 data:NDArray,
                 window:Window|None=None,
                 **tags):
-    # TODO: not_needed
+    # not_needed (could be useful though)
+    # no_work
+    # not_tested
     """Find a specific band and update it with data
 
     ..Note::
@@ -475,10 +456,6 @@ def update_band(src:DatasetWriter,
       Arbitrary number of keyword arguments that will be used to find
       the band to write into
     """
-    # not_needed (could be useful though)
-    # no_work
-    # not_tested
-    # usedin_both (could be)
     try:
         bidx = get_bidx(src=src, **tags)
     except BandSelectionNoMatchError as e:
@@ -491,13 +468,14 @@ def update_band(src:DatasetWriter,
     else:
         src.write(data, indexes=bidx, window=window)
 
-
 def export_to_tif(destination:str,
                   data:NDArray,
                   orig_profile:dict,
                   start=(0, 0),
                   **pparams):
-    # TODO: not_needed
+    # not_needed (could be useful though)
+    # no_work
+    # not_tested
     """Export a np.array to tif, only updating a window if data is smaller
 
     .. note::
@@ -518,10 +496,6 @@ def export_to_tif(destination:str,
     **pparams:
         further parameter to be added to the profile
     """
-    # not_needed (could be useful though)
-    # no_work
-    # not_tested
-    # usedin_both (pot.)
     profile = orig_profile.copy()
     # Note: we no longer update the size automatically as for Windows this is
     # not correct, pass height and width explicitly to update via pparams
@@ -536,9 +510,10 @@ def export_to_tif(destination:str,
     with rasterio.open(destination, "w", **profile) as dest:
         dest.write(data, window=Window(*start, *size), indexes=1)
 
-
 def project_to(source, reference, output=None, nodata=None)->str | None:
-    # TODO: not_needed
+    # not_needed (used in examples)
+    # no_work
+    # not_tested
     """Re-projects the source map into the coordinate system of a reference map
 
     Parameters
@@ -561,10 +536,6 @@ def project_to(source, reference, output=None, nodata=None)->str | None:
     str:
       The name of the file that hold the re-projected map
     """
-    # not_needed (used in examples)
-    # no_work
-    # not_tested
-    # usedin_both (potentially)
     with rio.open(reference) as ref:
         dst_crs = str(ref.profile['crs'])
     with rio.open(source) as src:
@@ -607,9 +578,10 @@ def project_to(source, reference, output=None, nodata=None)->str | None:
                 )
         return output
 
-
 def clip_to_bounds(source, reference, output=None):
-    # TODO: not_needed
+    # not_needed (used in example only)
+    # no_work
+    # not_tested
     """Clip raster to bounding box of reference raster
 
     Parameters
@@ -626,10 +598,6 @@ def clip_to_bounds(source, reference, output=None):
     str:
       The name of the file that holds clipped map
     """
-    # not_needed (used in example only)
-    # no_work
-    # not_tested
-    # usedin_both (potentially)
     if not check_crs_raster(source, reference):
         raise ValueError("Cannot clip by BBOX - projections are not the same.")
 
@@ -654,9 +622,10 @@ def clip_to_bounds(source, reference, output=None):
         dst.write(out_image)
     return output
 
-
 def _coregister_raster(source, reference, output=None):
-    # TODO: needed (Internal funciton only - move this to riogrande for now)
+    # is_needed (in tests only)
+    # needs_work (format doc)
+    # not_tested
     """Align raster to have identical resolution.
 
     Resolution will be calculated automatically from bounds and height/width of reference raster.
@@ -675,10 +644,6 @@ def _coregister_raster(source, reference, output=None):
     str:
       The name of the file that holds co-registered map
     """
-    # is_needed (in tests only)
-    # needs_work (format doc)
-    # not_tested
-    # usedin_both
 
     if not check_crs_raster(source, reference):
         raise ValueError("Cannot co-register sources - projections are not the same.")
@@ -720,9 +685,10 @@ def _coregister_raster(source, reference, output=None):
                     resampling=Resampling.nearest)
     return output
 
-
 def buffer_geometries_metric(geom_geoseries, buffer_meter, source_crs):
-    # TODO: not_needed
+    # is_needed (internal only)
+    # needs_work (docs)
+    # not_tested
     """ Applies a buffer to the geometries in GeoSeries given.
 
     ..Note: This function re-projects the GeoSeries to the respective UTM zone in order to use metric buffer and best
@@ -743,100 +709,16 @@ def buffer_geometries_metric(geom_geoseries, buffer_meter, source_crs):
     GeoSeries object:
       The buffered GeoSeries object
     """
-    # is_needed (internal only)
-    # needs_work (docs)
-    # not_tested
-    # usedin_both (io module)
     geom_utm = geom_geoseries.to_crs(geom_geoseries.estimate_utm_crs())
     geom_buff = geom_utm.buffer(buffer_meter,
                                 resolution=10, cap_style='round', join_style='round')
     geom_buff = geom_buff[geom_buff.area > 0]
     return geom_buff.to_crs(source_crs)
 
-
-def clip_to_ecoregion(source, shapefile, ecoregion_number, output=None, buffer_meter=None):
-    # TODO: not_needed
-    """Clip raster file to eco-region boundary (vector-data) for given eco-region number
-
-    Parameters
-    ----------
-    source: str
-      The path to the tif file you want to clip
-    shapefile: str
-      The path to the shapefile (.shp) with the eco-region polygons for clipping
-    ecoregion_number: int
-      The number of the respective eco-region to clip the source to
-    output: str (optional)
-      The path to write the clipped map to
-    buffer_meter: float, int
-      The buffer in meters to apply to the eco-region polygon before clipping.
-      Needs to be negative for reducing the polygon e.g. -1000
-
-    Returns
-    -------
-    str:
-      The name of the file that holds eco-region clipped map
-    """
-    # not_needed (used in example)
-    # no_work
-    # not_tested
-    # usedin_processing
-    if output is None:
-        output = outfile_suffix(source, "eco-clip")
-
-    with rasterio.open(source) as src:
-        bounds = src.bounds
-        bbox_geom = shbox(*bounds)
-
-    gdf = gpd.read_file(shapefile, bbox=bbox_geom)
-    gdf = gdf.loc[gdf.ecoregion == ecoregion_number]
-    geometry = gdf["geometry"]
-
-    # Buffer
-    # Note: before buffering cut shapefiler bigger than the raster to expanded smaler chunk to avoid long processing.
-    # Therfore the bbox of the raster plus the buffer distance (absolute) is used to avoid losing any areas later.
-    if buffer_meter is not None:
-        exp_bbox_geom = buffer_geometries_metric(gpd.GeoDataFrame(index=[0], crs='epsg:4326', geometry=[bbox_geom]),
-                                                 buffer_meter=abs(buffer_meter),
-                                                 source_crs=gdf.crs)
-        exp_geometry = gpd.clip(geometry, exp_bbox_geom, keep_geom_type=True)
-
-        geometry = buffer_geometries_metric(exp_geometry,
-                                            buffer_meter=buffer_meter,
-                                            source_crs=gdf.crs)
-
-    # Clip to bbox
-    geometry_clip = gpd.clip(geometry, bbox_geom, keep_geom_type=True)
-
-    # Make sure there are no Multilinestrings
-    geom_types = [t for t in geometry_clip.geom_type.unique()]
-    if len(geom_types) != 1:
-        if {'MultiPolygon', 'Polygon'} == set(geom_types):
-            geometry_clip = geometry_clip.explode(ignore_index=True)
-            if 'MultiPolygon' == geometry_clip.geom_type.unique():
-                raise ValueError("Issue in transforming MultiPolygons to Polygons using geopandas.explode")
-        else:
-            raise ValueError(f'Unvalid geometry types in clipping json {geom_types}')
-
-    # Clip files using shapely geometries as list and rasterio
-    cutline_shape = [geom for geom in geometry_clip]
-
-    with rio.open(source) as src:
-        out_array, out_transform = mask(src, cutline_shape, crop=False) # False to keep original dimension of raster
-        profile = src.profile.copy()
-
-        profile.update({"driver": "GTiff",
-                         "height": out_array.shape[1],
-                         "width": out_array.shape[2],
-                         "transform": out_transform})
-
-        with rio.open(output, "w", **profile) as dst:
-            dst.write(out_array)
-    return output
-
-
 def compress_tif(source, output:str|None=None, compression:str|None='lzw'):
-    # TODO: is_needed (technically not_needed but still present in some bash-scripts (Folder: scripts)
+    # is_needed
+    # needs_work (docs)
+    # is_tested
     # TODO: once this goes, we may also remove the outfile_suffix I believe (but double-check)
     """Compress tif file with LZW compression
 
@@ -855,10 +737,6 @@ def compress_tif(source, output:str|None=None, compression:str|None='lzw'):
     str:
       The name of the compressed file
     """
-    # is_needed
-    # needs_work (docs)
-    # is_tested
-    # usedin_processing (part of IO module, could be used in both)
     if compression is None:
         compression = 'none'
     overwrite = False
